Add Japan Honda Civic Custom RWD
This is an attempt at a linear torque mod. On this instance, it goes from 0x0000 to 0x7080 between the first and second values and stays at the max all the way through the table. Will this break everything and I'll need to go back and try gradually increasing it as per the normal method? Only time will tell.